pyproject.toml
[project]
name = "govuk-chat-search-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.2",
    "mcp[cli]>=1.9.2",
    "opensearch-py>=2.8.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
]

# Pyright LSP pick up uv venv directory
[tool.pyright]
venvPath = "."
venv = ".venv"
